+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/action-handlers/field/services/delete-field-action-handler.service.ts
@@ -1,0 +1,101 @@
+import { In } from 'typeorm';
+
+import { createWorkspaceMigrationRunnerActionHandler } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/interfaces/workspace-migration-runner-action-handler-service.interface';
+
+import { FieldMetadataEntity } from 'src/engine/metadata-modules/field-metadata/field-metadata.entity';
+import { findFlatFieldMetadataInFlatObjectMetadataMapsOrThrow } from 'src/engine/metadata-modules/flat-object-metadata-maps/utils/find-flat-field-metadata-in-flat-object-metadata-maps-or-throw.util';
+import { findFlatObjectMetadataInFlatObjectMetadataMapsOrThrow } from 'src/engine/metadata-modules/flat-object-metadata-maps/utils/find-flat-object-metadata-in-flat-object-metadata-maps-or-throw.util';
+import { findFlatObjectMetadataWithFlatFieldMapsInFlatObjectMetadataMapsOrThrow } from 'src/engine/metadata-modules/flat-object-metadata-maps/utils/find-flat-object-metadata-with-flat-field-maps-in-flat-object-metadata-maps-or-throw.util';
+import { type WorkspaceSchemaManagerService } from 'src/engine/twenty-orm/workspace-schema-manager/workspace-schema-manager.service';
+import { type DeleteFieldAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-field-action-v2';
+import { type WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
+import { generateColumnDefinitions } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/utils/generate-column-definitions.util';
+import { getWorkspaceSchemaContextForMigration } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/utils/get-workspace-schema-context-for-migration.util';
+import {
+  collectEnumOperationsForField,
+  EnumOperation,
+  executeBatchEnumOperations,
+} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/utils/workspace-schema-enum-operations.util';
+
+export class DeleteFieldActionHandlerService extends createWorkspaceMigrationRunnerActionHandler(
+  'delete_field',
+) {
+  constructor(
+    private readonly workspaceSchemaManagerService: WorkspaceSchemaManagerService,
+  ) {
+    super();
+  }
+
+  async executeForMetadata(
+    context: WorkspaceMigrationActionRunnerArgs<DeleteFieldAction>,
+  ) {
+    const { action, queryRunner } = context;
+    const fieldMetadataRepository =
+      queryRunner.manager.getRepository<FieldMetadataEntity>(
+        FieldMetadataEntity,
+      );
+
+    const { fieldMetadataId } = action;
+
+    await fieldMetadataRepository.delete({
+      id: In([fieldMetadataId]),
+    });
+  }
+
+  async executeForWorkspaceSchema(
+    context: WorkspaceMigrationActionRunnerArgs<DeleteFieldAction>,
+  ) {
+    const { action, queryRunner, flatObjectMetadataMaps, workspaceId } =
+      context;
+    const { objectMetadataId, fieldMetadataId } = action;
+
+    const flatObjectMetadataWithFlatFieldMaps =
+      findFlatObjectMetadataWithFlatFieldMapsInFlatObjectMetadataMapsOrThrow({
+        flatObjectMetadataMaps,
+        objectMetadataId,
+      });
+
+    const fieldMetadata = findFlatFieldMetadataInFlatObjectMetadataMapsOrThrow({
+      flatObjectMetadataMaps,
+      objectMetadataId,
+      fieldMetadataId,
+    });
+
+    const { schemaName, tableName } = getWorkspaceSchemaContextForMigration({
+      workspaceId,
+      flatObjectMetadata: flatObjectMetadataWithFlatFieldMaps,
+    });
+
+    const flatObjectMetadata =
+      findFlatObjectMetadataInFlatObjectMetadataMapsOrThrow({
+        flatObjectMetadataMaps,
+        objectMetadataId,
+      });
+
+    const columnDefinitions = generateColumnDefinitions({
+      flatFieldMetadata: fieldMetadata,
+      flatObjectMetadataWithoutFields: flatObjectMetadata,
+    });
+    const columnNamesToDrop = columnDefinitions.map((def) => def.name);
+
+    await this.workspaceSchemaManagerService.columnManager.dropColumns({
+      queryRunner,
+      schemaName,
+      tableName,
+      columnNames: columnNamesToDrop,
+    });
+
+    const enumOperations = collectEnumOperationsForField({
+      flatFieldMetadata: fieldMetadata,
+      tableName,
+      operation: EnumOperation.DROP,
+    });
+
+    await executeBatchEnumOperations({
+      enumOperations,
+      queryRunner,
+      schemaName,
+      workspaceSchemaManagerService: this.workspaceSchemaManagerService,
+    });
+  }
+}
