+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/action-handlers/index/services/delete-index-action-handler.service.ts
@@ -1,0 +1,20 @@
+import { createWorkspaceMigrationRunnerActionHandler } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/interfaces/workspace-migration-runner-action-handler-service.interface';
+
+import { type DeleteIndexAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-index-action-v2';
+import { type WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
+
+export class DeleteIndexActionHandlerService extends createWorkspaceMigrationRunnerActionHandler(
+  'delete_index',
+) {
+  async executeForMetadata(
+    _context: WorkspaceMigrationActionRunnerArgs<DeleteIndexAction>,
+  ): Promise<void> {
+    throw new Error('Not implemented');
+  }
+
+  async executeForWorkspaceSchema(
+    _context: WorkspaceMigrationActionRunnerArgs<DeleteIndexAction>,
+  ): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
