+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/action-handlers/object/services/delete-object-action-handler.service.ts
@@ -1,0 +1,87 @@
+import { createWorkspaceMigrationRunnerActionHandler } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/interfaces/workspace-migration-runner-action-handler-service.interface';
+
+import { type FlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type';
+import { isCompositeFlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/utils/is-composite-flat-field-metadata.util';
+import { isEnumFlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/utils/is-enum-flat-field-metadata.util';
+import { findFlatObjectMetadataWithFlatFieldMapsInFlatObjectMetadataMapsOrThrow } from 'src/engine/metadata-modules/flat-object-metadata-maps/utils/find-flat-object-metadata-with-flat-field-maps-in-flat-object-metadata-maps-or-throw.util';
+import { ObjectMetadataEntity } from 'src/engine/metadata-modules/object-metadata/object-metadata.entity';
+import { type WorkspaceSchemaManagerService } from 'src/engine/twenty-orm/workspace-schema-manager/workspace-schema-manager.service';
+import { type DeleteObjectAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-object-action-v2';
+import { type WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
+import { getWorkspaceSchemaContextForMigration } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/utils/get-workspace-schema-context-for-migration.util';
+import {
+  collectEnumOperationsForObject,
+  EnumOperation,
+  executeBatchEnumOperations,
+} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/utils/workspace-schema-enum-operations.util';
+
+export class DeleteObjectActionHandlerService extends createWorkspaceMigrationRunnerActionHandler(
+  'delete_object',
+) {
+  constructor(
+    private readonly workspaceSchemaManagerService: WorkspaceSchemaManagerService,
+  ) {
+    super();
+  }
+
+  async executeForMetadata(
+    context: WorkspaceMigrationActionRunnerArgs<DeleteObjectAction>,
+  ): Promise<void> {
+    const { action, queryRunner } = context;
+    const { objectMetadataId } = action;
+
+    const objectMetadataRepository =
+      queryRunner.manager.getRepository<ObjectMetadataEntity>(
+        ObjectMetadataEntity,
+      );
+
+    await objectMetadataRepository.delete(objectMetadataId);
+  }
+
+  async executeForWorkspaceSchema(
+    context: WorkspaceMigrationActionRunnerArgs<DeleteObjectAction>,
+  ): Promise<void> {
+    const { action, queryRunner, flatObjectMetadataMaps, workspaceId } =
+      context;
+    const { objectMetadataId } = action;
+
+    const flatObjectMetadataWithFlatFieldMaps =
+      findFlatObjectMetadataWithFlatFieldMapsInFlatObjectMetadataMapsOrThrow({
+        flatObjectMetadataMaps,
+        objectMetadataId,
+      });
+
+    const { schemaName, tableName } = getWorkspaceSchemaContextForMigration({
+      workspaceId,
+      flatObjectMetadata: flatObjectMetadataWithFlatFieldMaps,
+    });
+
+    await this.workspaceSchemaManagerService.tableManager.dropTable({
+      queryRunner,
+      schemaName,
+      tableName,
+    });
+
+    const enumOrCompositeFlatFieldMetadatas = Object.values(
+      flatObjectMetadataWithFlatFieldMaps.fieldsById,
+    )
+      .filter((field): field is FlatFieldMetadata => field != null)
+      .filter(
+        (field) =>
+          isEnumFlatFieldMetadata(field) || isCompositeFlatFieldMetadata(field),
+      );
+
+    const enumOperations = collectEnumOperationsForObject({
+      flatFieldMetadatas: enumOrCompositeFlatFieldMetadatas,
+      tableName,
+      operation: EnumOperation.DROP,
+    });
+
+    await executeBatchEnumOperations({
+      enumOperations,
+      queryRunner,
+      schemaName,
+      workspaceSchemaManagerService: this.workspaceSchemaManagerService,
+    });
+  }
+}
