+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/exceptions/workspace-schema-migration.exception.ts
@@ -1,0 +1,10 @@
+import { CustomException } from 'src/utils/custom-exception';
+
+export class WorkspaceSchemaMigrationException extends CustomException<WorkspaceSchemaMigrationExceptionCode> {}
+
+export enum WorkspaceSchemaMigrationExceptionCode {
+  ENUM_OPERATION_FAILED = 'ENUM_OPERATION_FAILED',
+  UNSUPPORTED_COMPOSITE_COLUMN_TYPE = 'UNSUPPORTED_COMPOSITE_COLUMN_TYPE',
+  NOT_SUPPORTED = 'NOT_SUPPORTED',
+  INVALID_ACTION_TYPE = 'INVALID_ACTION_TYPE',
+}
