--- packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/runner-method-for-action-type.ts
@@ -1,17 +1,0 @@
-import { type ConvertActionTypeToCamelCase } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/convert-action-type-to-camel-case.type';
-import {
-  type WorkspaceMigrationActionTypeV2,
-  type WorkspaceMigrationActionV2,
-} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-action-common-v2';
-import { type WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
-
-export type RunnerMethodForActionType<
-  TAction extends WorkspaceMigrationActionTypeV2,
-  TRunner extends 'metadata' | 'schema',
-> = {
-  [P in TAction as `run${Capitalize<ConvertActionTypeToCamelCase<P>>}${Capitalize<TRunner>}Migration`]: (
-    arg: WorkspaceMigrationActionRunnerArgs<
-      Extract<WorkspaceMigrationActionV2, { type: P }>
-    >,
-  ) => Promise<void>;
-};
