+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type.ts
@@ -9,4 +9,5 @@ export type WorkspaceMigrationActionRunnerArgs<
   queryRunner: QueryRunner;
   action: T;
   flatObjectMetadataMaps: FlatObjectMetadataMaps;
+  workspaceId: string;
 };
