--- packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-runner-args.type.ts
@@ -1,10 +1,0 @@
-import { type QueryRunner } from 'typeorm';
-
-import { type FlatObjectMetadataMaps } from 'src/engine/metadata-modules/flat-object-metadata-maps/types/flat-object-metadata-maps.type';
-import { type WorkspaceMigrationActionV2 } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-action-common-v2';
-
-export type WorkspaceMigrationRunnerArgs = {
-  action: WorkspaceMigrationActionV2;
-  queryRunner: QueryRunner;
-  flatObjectMetadataMaps: FlatObjectMetadataMaps;
-};
