+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/utils/get-workspace-schema-context-for-migration.util.ts
@@ -1,0 +1,24 @@
+import {
+  type FlatObjectMetadata,
+  type FlatObjectMetadataWithoutFields,
+} from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata.type';
+import { computeObjectTargetTable } from 'src/engine/utils/compute-object-target-table.util';
+import { getWorkspaceSchemaName } from 'src/engine/workspace-datasource/utils/get-workspace-schema-name.util';
+
+export type WorkspaceSchemaContextForMigration = {
+  schemaName: string;
+  tableName: string;
+};
+
+export const getWorkspaceSchemaContextForMigration = ({
+  workspaceId,
+  flatObjectMetadata,
+}: {
+  workspaceId: string;
+  flatObjectMetadata: FlatObjectMetadata | FlatObjectMetadataWithoutFields;
+}): WorkspaceSchemaContextForMigration => {
+  return {
+    schemaName: getWorkspaceSchemaName(workspaceId),
+    tableName: computeObjectTargetTable(flatObjectMetadata),
+  };
+};
