--- packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-metadata-migration-runner/workspace-metadata-field-action-runner.service.ts
@@ -1,63 +1,0 @@
-import { In } from 'typeorm';
-
-import { FieldMetadataEntity } from 'src/engine/metadata-modules/field-metadata/field-metadata.entity';
-import {
-  type CreateFieldAction,
-  type DeleteFieldAction,
-  type UpdateFieldAction,
-  type WorkspaceMigrationFieldActionTypeV2,
-} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-field-action-v2';
-import { type RunnerMethodForActionType } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/runner-method-for-action-type';
-import { type WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
-import { fromWorkspaceMigrationUpdateActionToPartialEntity } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/utils/from-workspace-migration-update-action-to-partial-field-or-object-entity.util';
-
-export class WorkspaceMetadataFieldActionRunnerService
-  implements
-    RunnerMethodForActionType<WorkspaceMigrationFieldActionTypeV2, 'metadata'>
-{
-  runDeleteFieldMetadataMigration = async ({
-    action,
-    queryRunner,
-  }: WorkspaceMigrationActionRunnerArgs<DeleteFieldAction>) => {
-    const fieldMetadataRepository =
-      queryRunner.manager.getRepository<FieldMetadataEntity>(
-        FieldMetadataEntity,
-      );
-
-    const { fieldMetadataId } = action;
-
-    await fieldMetadataRepository.delete({
-      id: In([fieldMetadataId]),
-    });
-  };
-
-  runCreateFieldMetadataMigration = async ({
-    action,
-    queryRunner,
-  }: WorkspaceMigrationActionRunnerArgs<CreateFieldAction>) => {
-    const fieldMetadataRepository =
-      queryRunner.manager.getRepository<FieldMetadataEntity>(
-        FieldMetadataEntity,
-      );
-
-    const { flatFieldMetadata } = action;
-
-    await fieldMetadataRepository.save(flatFieldMetadata);
-  };
-  runUpdateFieldMetadataMigration = async ({
-    action,
-    queryRunner,
-  }: WorkspaceMigrationActionRunnerArgs<UpdateFieldAction>) => {
-    const fieldMetadataRepository =
-      queryRunner.manager.getRepository<FieldMetadataEntity>(
-        FieldMetadataEntity,
-      );
-
-    const { fieldMetadataId } = action;
-
-    await fieldMetadataRepository.update(
-      fieldMetadataId,
-      fromWorkspaceMigrationUpdateActionToPartialEntity(action),
-    );
-  };
-}
