--- packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/action-handlers/object/services/create-object-action-handler.service.ts
@@ -1,72 +1,0 @@
-import { createWorkspaceMigrationActionHandler } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/core/interfaces/workspace-migration-action-service.interface';
-
-import { type FlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type';
-import { isCompositeFlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/utils/is-composite-flat-field-metadata.util';
-import { isEnumFlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/utils/is-enum-flat-field-metadata.util';
-import { type WorkspaceSchemaManagerService } from 'src/engine/twenty-orm/workspace-schema-manager/workspace-schema-manager.service';
-import { computeObjectTargetTable } from 'src/engine/utils/compute-object-target-table.util';
-import { getWorkspaceSchemaName } from 'src/engine/workspace-datasource/utils/get-workspace-schema-name.util';
-import { type CreateObjectAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-object-action-v2';
-import { type WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
-import { generateColumnDefinitions } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/generate-column-definitions.util';
-import {
-  collectEnumOperationsForObject,
-  EnumOperation,
-  executeBatchEnumOperations,
-} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/workspace-schema-enum-operations.util';
-
-export class CreateObjectActionHandlerService extends createWorkspaceMigrationActionHandler(
-  'create_object',
-) {
-  constructor(
-    private readonly workspaceSchemaManagerService: WorkspaceSchemaManagerService,
-  ) {
-    super();
-  }
-
-  async execute(
-    context: WorkspaceMigrationActionRunnerArgs<CreateObjectAction>,
-  ): Promise<void> {
-    const { action, queryRunner } = context;
-    const { flatObjectMetadataWithoutFields, createFieldActions } = action;
-
-    const flatObjectMetadata = flatObjectMetadataWithoutFields;
-    const schemaName = getWorkspaceSchemaName(flatObjectMetadata.workspaceId);
-    const tableName = computeObjectTargetTable(flatObjectMetadata);
-
-    const columnDefinitions = createFieldActions.flatMap((createFieldAction) =>
-      generateColumnDefinitions({
-        flatFieldMetadata: createFieldAction.flatFieldMetadata,
-        flatObjectMetadataWithoutFields: flatObjectMetadataWithoutFields,
-      }),
-    );
-
-    const enumOrCompositeFlatFieldMetadatas = createFieldActions
-      .map((createFieldAction) => createFieldAction.flatFieldMetadata)
-      .filter((field): field is FlatFieldMetadata => field != null)
-      .filter(
-        (field) =>
-          isEnumFlatFieldMetadata(field) || isCompositeFlatFieldMetadata(field),
-      );
-
-    const enumOperations = collectEnumOperationsForObject({
-      flatFieldMetadatas: enumOrCompositeFlatFieldMetadatas,
-      tableName,
-      operation: EnumOperation.CREATE,
-    });
-
-    await executeBatchEnumOperations({
-      enumOperations,
-      queryRunner,
-      schemaName,
-      workspaceSchemaManagerService: this.workspaceSchemaManagerService,
-    });
-
-    await this.workspaceSchemaManagerService.tableManager.createTable({
-      queryRunner,
-      schemaName,
-      tableName,
-      columnDefinitions,
-    });
-  }
-}
