--- packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/action-handlers/object/services/delete-object-action-handler.service.ts
@@ -1,65 +1,0 @@
-import { createWorkspaceMigrationActionHandler } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/core/interfaces/workspace-migration-action-service.interface';
-
-import { type FlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata.type';
-import { isCompositeFlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/utils/is-composite-flat-field-metadata.util';
-import { isEnumFlatFieldMetadata } from 'src/engine/metadata-modules/flat-field-metadata/utils/is-enum-flat-field-metadata.util';
-import { type WorkspaceSchemaManagerService } from 'src/engine/twenty-orm/workspace-schema-manager/workspace-schema-manager.service';
-import { type DeleteObjectAction } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/types/workspace-migration-object-action-v2';
-import { type WorkspaceMigrationActionRunnerArgs } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/types/workspace-migration-action-runner-args.type';
-import { prepareWorkspaceSchemaContext } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/workspace-schema-context.util';
-import {
-  collectEnumOperationsForObject,
-  EnumOperation,
-  executeBatchEnumOperations,
-} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/utils/workspace-schema-enum-operations.util';
-
-export class DeleteObjectActionHandlerService extends createWorkspaceMigrationActionHandler(
-  'delete_object',
-) {
-  constructor(
-    private readonly workspaceSchemaManagerService: WorkspaceSchemaManagerService,
-  ) {
-    super();
-  }
-
-  async execute(
-    context: WorkspaceMigrationActionRunnerArgs<DeleteObjectAction>,
-  ): Promise<void> {
-    const { action, queryRunner, flatObjectMetadataMaps } = context;
-    const { objectMetadataId } = action;
-
-    const { schemaName, tableName, flatObjectMetadataWithFlatFieldMaps } =
-      prepareWorkspaceSchemaContext({
-        flatObjectMetadataMaps,
-        objectMetadataId,
-      });
-
-    await this.workspaceSchemaManagerService.tableManager.dropTable({
-      queryRunner,
-      schemaName,
-      tableName,
-    });
-
-    const enumOrCompositeFlatFieldMetadatas = Object.values(
-      flatObjectMetadataWithFlatFieldMaps.fieldsById,
-    )
-      .filter((field): field is FlatFieldMetadata => field != null)
-      .filter(
-        (field) =>
-          isEnumFlatFieldMetadata(field) || isCompositeFlatFieldMetadata(field),
-      );
-
-    const enumOperations = collectEnumOperationsForObject({
-      flatFieldMetadatas: enumOrCompositeFlatFieldMetadatas,
-      tableName,
-      operation: EnumOperation.DROP,
-    });
-
-    await executeBatchEnumOperations({
-      enumOperations,
-      queryRunner,
-      schemaName,
-      workspaceSchemaManagerService: this.workspaceSchemaManagerService,
-    });
-  }
-}
