--- packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/core/workspace-schema-migration-runner-core.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DiscoveryModule } from '@nestjs/core';
-
-import { WorkspaceSchemaActionRegistryService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-schema-migration-runner/core/registry/workspace-schema-action-registry.service';
-
-@Module({
-  imports: [DiscoveryModule],
-  providers: [WorkspaceSchemaActionRegistryService],
-  exports: [WorkspaceSchemaActionRegistryService],
-})
-export class WorkspaceSchemaMigrationRunnerCoreModule {}
