+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-v2.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
-import { WorkspaceMigrationBuilderV2Module } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/services/workspace-migration-builder-v2.module';
+import { WorkspaceMigrationValidateBuildAndRunService } from 'src/engine/workspace-manager/workspace-migration-v2/services/workspace-migration-validate-build-and-run-service';
+import { WorkspaceMigrationBuilderV2Module } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/workspace-migration-builder-v2.module';
 import { WorkspaceMigrationRunnerV2Module } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-runner-v2/workspace-migration-runner-v2.module';
-import { WorkspaceMigrationValidateBuildAndRunService } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-validate-build-and-run-service';
 
 @Module({
   imports: [
@@ -10,10 +10,6 @@ import { WorkspaceMigrationValidateBuildAndRunService } from 'src/engine/workspa
     WorkspaceMigrationRunnerV2Module,
   ],
   providers: [WorkspaceMigrationValidateBuildAndRunService],
-  exports: [
-    WorkspaceMigrationRunnerV2Module,
-    WorkspaceMigrationBuilderV2Module,
-    WorkspaceMigrationValidateBuildAndRunService,
-  ],
+  exports: [WorkspaceMigrationValidateBuildAndRunService],
 })
 export class WorkspaceMigrationV2Module {}
